ruby symlink.

+++ ruby19/recipes/default.rb
@@ -16,7 +16,9 @@
 	end
 end
 
-bash "make symlik" do 
+bash "make symlik" do
+    only_if { File.exists?('/usr/bin/gem1.9')}
+    not_if { File.exists?('/usr/bin/gem')}
 	code <<-EOC
 		rm /usr/bin/ruby
 		ln -s /usr/bin/ruby1.9 /usr/bin/ruby
